move test with time.Sleep into synctest file
